requirements.txt
netmiko==4.3.0
fastapi==0.111.0
uvicorn==0.30.1
pyyaml==6.0.1
networkx==3.3